Add htq.status() and endpoint for request status
This is a lightweight call to get the status of the request for
determining when the response is ready. For this reason, the response
endpoint can now be accessed after the request has been queued and will
block until the response is ready.

The recommended approach for clients is to poll the status endpoint
until it is `success`, `error`, or `timeout`.

Signed-off-by: Byron Ruth <b@devel.io>